+++ App/Componentes/seccionCartelera.mjs
@@ -1,11 +1,9 @@
-import { CrearImg, CrearDiv } from "./funciones.mjs";
-import { Carrusel } from "./cargaInicio.mjs";
+import { CrearImg,CrearDiv } from "./funciones.mjs";
 const carteleraApp = () => {
     const $cartelera = document.getElementById("id-cartelera");
     const $secInicio = document.getElementById("sec-inicio");
     const $secCartelera = document.getElementById("sec-cartelera");
     const $secCombo = document.getElementById("sec-combo");
-    const $divMain = document.getElementById("page-car");
 
     $cartelera.addEventListener("click", () => {
         $secInicio.style.display = "none";
@@ -13,7 +11,56 @@ const carteleraApp = () => {
         $secCombo.style.display = "none";
     });
 
-    Carrusel();
+    let $divMain = document.createElement("div");
+    let $divImg = document.createElement("div");
+    $divImg.classList.add("contenedorImagenCartelera");
+
+    $divMain.classList.add("page");
+    $secCartelera.appendChild($divMain);
+    $divMain.appendChild($divImg);
+
+    let $btnA = document.createElement("button");
+    let $btnS = document.createElement("button");
+    $btnA.classList.add("btnA");
+    $btnS.classList.add("btnS");
+    $btnA.textContent = "<";
+    $btnS.textContent = ">";
+    $divImg.appendChild($btnA);
+    $divImg.appendChild($btnS);
+
+
+    $divImg.appendChild(CrearImg("carousel-image", "./img_peliculas/img_1/kong_1.jpg"));
+
+    let i = 0;
+    const $imagenes = [
+        "./img_peliculas/img_1/kong_1.jpg",
+        "./img_peliculas/img_1/michi_1.jpg",
+        "./img_peliculas/img_1/panda_1.jpg"
+    ];
+
+    const $image = document.querySelector(".carousel-image");
+    const mostrarImagen = (index) => {
+        $image.src = $imagenes[index];
+    };
+
+    const siguienteImagen = () => {
+        i = (i + 1) % $imagenes.length;
+        mostrarImagen(i);
+    };
+
+    const imagenAnterior = () => {
+        i = (i - 1 + $imagenes.length) % $imagenes.length;
+        mostrarImagen(i);
+    };
+
+    $btnA.addEventListener("click", () => {
+        siguienteImagen();
+    });
+    $btnS.addEventListener("click", () => {
+        imagenAnterior();
+    });
+
+    setInterval(siguienteImagen, 3000);
 
     let $divPoster = document.createElement("div");
     $divPoster.classList.add("ContenidoPoster");
@@ -23,7 +70,7 @@ const carteleraApp = () => {
     let $li1 = document.createElement("li");
     let $li2 = document.createElement("li");
     let $li3 = document.createElement("li");
-
+    
     $li1.appendChild(CrearImg("img_Poster1", "./img_peliculas/img_2/kong_2.jpg"));
     $li2.appendChild(CrearImg("img_Poster2", "./img_peliculas/img_2/michi_2.jpg"));
     $li3.appendChild(CrearImg("img_Poster3", "./img_peliculas/img_2/panda_2.jpg"));
@@ -33,7 +80,7 @@ const carteleraApp = () => {
     $li3.appendChild(CrearDiv("Luchito - animada"));
 
     let $ul = document.createElement("ul");
-    $ul.setAttribute("id", "GaleriaPoster");
+    $ul.setAttribute("id","GaleriaPoster");
     $ul.appendChild($li1);
     $ul.appendChild($li2);
     $ul.appendChild($li3);
