+++ App/Componentes/seccionCombos.mjs
@@ -1,11 +1,9 @@
 import { CrearImg } from "./funciones.mjs";
-import { Carrusel } from "./cargaInicio.mjs";
 const combosApp = () => {
     const $combos = document.getElementById("id-combos");
     const $secInicio = document.getElementById("sec-inicio");
     const $secCartelera = document.getElementById("sec-cartelera");
     const $secCombo = document.getElementById("sec-combo");
-    const $divMain = document.getElementById("page-com");
 
     $combos.addEventListener("click", () => {
         $secInicio.style.display = "none";
@@ -13,17 +11,63 @@ const combosApp = () => {
         $secCombo.style.display = "block";
     });
 
-    Carrusel();
+    let $divContenido = document.createElement("div");
+    let $divImg = document.createElement("div");
+    $divImg.classList.add("contenedorImagenCartelera");
+
+    $divContenido.classList.add("page");
+    $secCombo.appendChild($divContenido);
+    $divContenido.appendChild($divImg);
+
+    let $btnA = document.createElement("button");
+    let $btnS = document.createElement("button");
+    $btnA.classList.add("btnA");
+    $btnS.classList.add("btnS");
+    $btnA.textContent = "<";
+    $btnS.textContent = ">";
+    $divImg.appendChild($btnA);
+    $divImg.appendChild($btnS);
+
+    $divImg.appendChild(CrearImg("carousel-image", "./img_peliculas/img_1/kong_1.jpg"));
+    let i = 0;
+    const $imagenes = [ // No andan ni idea
+        "./img_peliculas/img_1/kong_1.jpg",
+        "./img_peliculas/img_1/michi_1.jpg",
+        "./img_peliculas/img_1/panda_1.jpg"
+    ];
+    const $image = document.querySelector(".carousel-image");
+    const mostrarImagen = (index) => {
+        $image.src = $imagenes[index];
+    };
+
+    const siguienteImagen = () => {
+        i = (i + 1) % $imagenes.length;
+        mostrarImagen(i);
+    };
+
+    const imagenAnterior = () => {
+        i = (i - 1 + $imagenes.length) % $imagenes.length;
+        mostrarImagen(i);
+    };
+
+    $btnA.addEventListener("click", () => {
+        siguienteImagen();
+    });
+    $btnS.addEventListener("click", () => {
+        imagenAnterior();
+    });
+
+    setInterval(siguienteImagen, 3000);
 
     let $divCombos = document.createElement("div");
     let $hCombo = document.createElement("h2");
     $hCombo.textContent = "Combos";
 
     $divCombos.appendChild($hCombo);
-    $divMain.appendChild($divCombos);
+    $divContenido.appendChild($divCombos);
 
     let $divCombosC = document.createElement("div");
-    $divMain.appendChild($divCombosC);
+    $divContenido.appendChild($divCombosC);
 
     const CrearDivImg = (nombreClass, rutaImg, contenidoText) => {
         let $divComb = document.createElement("div");
